Renamed pkg dir to internal
